pyproject.toml
[project]
name = "gradescope-ics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "flask>=3.1.0",
    "ics>=0.7.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
